数据库数据格式要注意， 更新了wm position数据的关联

+++ src/main/java/com/zy/in/dao/InMapper.java
@@ -8,7 +8,7 @@ import org.apache.ibatis.annotations.Param;
 import com.zy.in.entities.In;
 
 public interface InMapper {
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(In record);
